Separate IconCollectionSearchTool and IconInformationModal
